requirements.txt
bcrypt==5.0.0
flask==3.1.3
flask-cors==6.0.2
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
rasterio
numpy
scikit-learn